Preserve file extensions in temporary files
This is needed to work with rio and potentially other functions that
inspect the file extension and do varying things based on that.  Functions
that vary based on the *path* are a potentially more difficult, but I've
added stub support for that (need to pass it along in the interface still).