Use by default the openssl package keygen
This is an issue only for old operating systems that do not
have a ssh key that want to use a ssh key from the command line
but not generate one from the command line - that's a fairly small
intersection!  The ability to generate one is still there though
so hopefully this can be made to work reasonably well.